refactor: 기존타입 bigdicmal로 변경

+++ src/main/java/com/nhnacademy/illuwa/domain/coupons/dto/couponPolicy/CouponPolicyCreateResponse.java
@@ -9,6 +9,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -20,10 +21,10 @@ public class CouponPolicyCreateResponse {
    private Long id;
    private String code;
    private CouponStatus status;
-   private Integer minOrderAmount;
-   private Integer discountAmount;
-   private Integer discountPercent;
-   private Integer maxDiscountAmount;
+   private BigDecimal minOrderAmount;
+   private BigDecimal discountAmount;
+   private BigDecimal discountPercent;
+   private BigDecimal maxDiscountAmount;
    private LocalDateTime createdAt;
 
     public static CouponPolicyCreateResponse fromEntity(CouponPolicy policy) {
