+++ src/main/java/com/nhnacademy/illuwa/domain/coupons/dto/couponPolicy/CouponPolicyUpdateRequest.java
@@ -6,18 +6,20 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
 public class CouponPolicyUpdateRequest {
-    private Integer minOrderAmount; // 최소 주문 금액
+    private BigDecimal minOrderAmount; // 최소 주문 금액
 
-    private Integer discountAmount; // 할인 금액
+    private BigDecimal discountAmount; // 할인 금액
 
-    private Integer discountPercent; // 할인 퍼센트
+    private BigDecimal discountPercent; // 할인 퍼센트
 
-    private Integer maxDiscountAmount; // 최대 할인 금액
+    private BigDecimal maxDiscountAmount; // 최대 할인 금액
 
     private CouponStatus status; // 상태 여부 수정
